+++ src/constellation/api_devices_create.go
@@ -12,18 +12,19 @@ type DeviceCreateRequestJSON struct {
 	DeviceName string `json:"deviceName",validate:"required,min=3,max=32,alphanum"`
 	IP string `json:"ip",validate:"required,ipv4"`
 	PublicKey string `json:"publicKey",omitempty`
-	
+
 	// for devices only
 	Nickname string `json:"nickname",validate:"max=32,alphanum",omitempty`
 	Invisible bool `json:"invisible",omitempty`
-	
+
 	// for lighthouse only
 	IsLighthouse bool `json:"isLighthouse",omitempty`
+	IsCosmosNode bool `json:"isCosmosNode",omitempty`
 	IsRelay bool `json:"isRelay",omitempty`
 	IsExitNode bool `json:"isExitNode",omitempty`
 	PublicHostname string `json:"PublicHostname",omitempty`
 	Port string `json:"port",omitempty`
-	
+
 }
 
 func DeviceCreate(w http.ResponseWriter, req *http.Request) {
@@ -37,7 +38,7 @@ func DeviceCreate(w http.ResponseWriter, req *http.Request) {
 			return 
 		}
 
-		if !utils.FBL.LValid {
+		if !utils.FBL.LValid && !utils.FBL.IsCosmosNode {
 			utils.Error("ConstellationDeviceCreation: No valid licence found to use Constellation.", nil)
 			utils.HTTPError(w, "Device Creation Error: No valid licence found to use Constellation.",
 				http.StatusInternalServerError, "DC001")
@@ -99,6 +100,47 @@ func DeviceCreate(w http.ResponseWriter, req *http.Request) {
 				return
 			}
 
+			// Cosmos nodes are also lighthouses
+			if request.IsCosmosNode {
+				request.IsLighthouse = true
+			}
+
+			// Check cosmos node and devices limit
+			if request.IsCosmosNode {
+				totalClientLimit := 10 * int64(utils.GetNumberUsers())
+
+				count, errCount := c.CountDocuments(nil, map[string]interface{}{
+					"IsCosmosNode": true,
+					"Blocked": false,
+				})
+				if errCount != nil {
+					utils.Error("DeviceCreation: Error while counting cosmos nodes", errCount)
+					utils.HTTPError(w, "Device Creation Error", http.StatusInternalServerError, "DC009")
+					return
+				}
+
+				countDevices, errCountDevices := c.CountDocuments(nil, map[string]interface{}{
+					"Blocked": false,
+				})
+				if errCountDevices != nil {
+					utils.Error("DeviceCreation: Error while counting devices", errCountDevices)
+					utils.HTTPError(w, "Device Creation Error", http.StatusInternalServerError, "DC011")
+					return
+				}
+
+				if countDevices >= totalClientLimit {
+					utils.Error("DeviceCreation: Device limit reached", nil)
+					utils.HTTPError(w, "Device limit reached", http.StatusConflict, "DC012")
+					return
+				}
+
+				if count >= int64(utils.GetNumberCosmosNode()) {
+					utils.Error("DeviceCreation: Cosmos node limit reached", nil)
+					utils.HTTPError(w, "Cosmos node limit reached", http.StatusConflict, "DC010")
+					return
+				}
+			}
+
 			if request.IsLighthouse && request.Nickname != "" {
 				utils.Error("DeviceCreation: Lighthouse cannot belong to a user", nil)
 				utils.HTTPError(w, "Device Creation Error: Lighthouse cannot have a nickname",
@@ -119,6 +161,7 @@ func DeviceCreate(w http.ResponseWriter, req *http.Request) {
 				"PublicKey": key,
 				"IP": request.IP,
 				"IsLighthouse": request.IsLighthouse,
+				"IsCosmosNode": request.IsCosmosNode,
 				"IsRelay": request.IsLighthouse && request.IsRelay,
 				"IsExitNode": request.IsLighthouse && request.IsExitNode,
 				"PublicHostname": request.PublicHostname,
@@ -157,6 +200,7 @@ func DeviceCreate(w http.ResponseWriter, req *http.Request) {
 				PublicKey: key,
 				IP: request.IP,
 				IsLighthouse: request.IsLighthouse,
+				IsCosmosNode: request.IsCosmosNode,
 				IsRelay: request.IsLighthouse && request.IsRelay,
 				IsExitNode: request.IsLighthouse && request.IsExitNode,
 				PublicHostname: request.PublicHostname,
@@ -195,6 +239,7 @@ func DeviceCreate(w http.ResponseWriter, req *http.Request) {
 					"Config": configYml,
 					"CA": capki,
 					"IsLighthouse": request.IsLighthouse,
+					"IsCosmosNode": request.IsCosmosNode,
 					"IsRelay": request.IsLighthouse && request.IsRelay,
 					"IsExitNode": request.IsLighthouse && request.IsExitNode,
 					"PublicHostname": request.PublicHostname,